[Refactor] seperate scheduler from service
