+++ src/main/java/Dino/Duett/domain/message/service/MessageService.java
@@ -19,12 +19,9 @@ import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.*;
 
 @Service
@@ -110,12 +107,5 @@ public class MessageService {
         //messageRepository.deleteAllByIdIn(deleteMessageIds);
 
         return MessageDeleteResponse.of(deleteMessageIds);
-    }
-
-    @Scheduled(cron = "0 0 4 * * ?", zone = "Asia/Seoul")
-    @Transactional
-    public void deleteOldMessagesScheduler() {
-        LocalDate cutoffDate = LocalDateTime.now().toLocalDate().minusWeeks(2);
-        messageRepository.deleteByCreatedDateBefore(cutoffDate.atStartOfDay());
     }
 }
